update CRUD berdasarkan user yang aktif

+++ app/Http/Controllers/Todo/TodoController.php
@@ -2,11 +2,12 @@
 
 namespace App\Http\Controllers\Todo;
 
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
-
 use App\Models\Todo;
 use App\Models\User;
+
+use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 
 class TodoController extends Controller
 {
@@ -20,9 +21,9 @@ class TodoController extends Controller
         $max_data = 3;
 
         if(request('search')) {
-            $data = Todo::where('task', 'like', '%' .  request('search') . '%')->paginate($max_data)->withQueryString();
+            $data = Todo::where('task', 'like', '%' .  request('search') . '%')->where('user_id', Auth::user()->id)->orderBy('task', 'asc')->paginate($max_data)->withQueryString();
         } else {
-            $data = Todo::orderBy('task', 'asc')->paginate($max_data);
+            $data = Todo::orderBy('task', 'asc')->where('user_id', Auth::user()->id)->paginate($max_data);
         }
 
         // return view("todo.app", ['data' => $data]); or
@@ -51,7 +52,8 @@ class TodoController extends Controller
         ]);
 
         $data = [
-            'task' => $request->input('task')
+            'task' => $request->input('task'),
+            'user_id' => Auth::user()->id
         ];
 
         Todo::create($data);
@@ -90,10 +92,10 @@ class TodoController extends Controller
 
         $data = [
             'task'      => $request->input('task'),
-            'is_done'   => $request->input('is_done')
+            'is_done'   => $request->input('is_done'),
         ];
 
-        Todo::where('id', $id)->update($data);
+        Todo::where('id', $id)->where('user_id', Auth::user()->id)->update($data);
         return redirect()->route('todo')->with('success', 'Data berhasil di update');
     }
 
@@ -102,7 +104,7 @@ class TodoController extends Controller
      */
     public function destroy(string $id)
     {
-        Todo::where('id', $id)->delete();
+        Todo::where('id', $id)->where('user_id', Auth::user()->id)->delete();
         return redirect()->route('todo')->with('success', 'Data berhasil di hapus');
     }
 }
